+++ src/Entity/Volume.php
@@ -267,7 +267,7 @@ class Volume extends AbstractVolumeSection implements EntityIdentifierInterface
     #[Groups(
         [
             AppConstants::APP_CONST['normalizationContext']['groups']['volume']['item']['read'][0],
-            #AppConstants::APP_CONST['normalizationContext']['groups']['volume']['collection']['read'][0]
+            AppConstants::APP_CONST['normalizationContext']['groups']['volume']['collection']['read'][0]
         ]
 
     )]